pyproject.toml
[project]
name = "incubation"
version = "0.1.0"
description = "..."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "langchain-model-profiles>=0.0.5",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-ollama>=1.0.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "langchain>=1.1.0",
    "langgraph-cli>=0.1.0",
]
